add antispam test
